Indirectly import MongoDB libraries
Builds would fail since SvelteKit would keep trying to include the
unused `mongo-client-encryption`, which would fail upon building itself.

I tried looking for ways to install this `mongo-client-encryption` but
to no avail. So I tried to cut the `require` statement out manually by
editing `node_modules` by hand. An error message came out warning that
we couldn't use the previous import statement and gave a suggestion:

    import pkg from 'mongodb';
    const {ObjectId, MongoClient} = pkg;

Previously it looked like this:

    const { ObjectId, MongoClient } from "mongodb";

I understand that when you import stuff on the server-side, you usually
want to use `require` since that's where it lives (and not on the
client-side). However, this would actually fail since `require` doesn't
seem to live on the server-side anymore. I think this is a consequence
of using SvelteKit, which aims to be deployable on serverless platforms.