Manually patch MongoDB to avoid build failure
Context: c5343b637679748859a0b9f0286949a2cdd94ec3

I couldn't figure out how to prevent the MongoDB import from requiring
`mongodb-client-encryption`, so I put the line in a try-catch to prevent
it from working.

I used the fix from https://stackoverflow.com/a/62567504 which lets us
produce a patch that can be reapplied whenever we update dependencies.

This is kinda ugly, but the only alternative is to open an issue on
MongoDB's issue tracker. We'll have to come back to this later.