Assign unique debt ID on creation
Previously we let MongoDB assign the ID upon insertion. This means that
there's a split second where the actual debt data does not have an ID,
which would mess with the keyed `each` block in routes/index.svelte.

This time we ask for a unique ID and assign it to the new debt directly
before it gets inserted into the database.