Make api/debt/[id].json.ts PATCH method parameters more explicit
Previously, we were able to set the `is_paid` field of a debt by using
the endpoint like this:

    await fetch(`api/debt/${some_debt_id}.json?is_paid`);

To set `is_paid` to false, we omit the `?is_paid` part of the URL.

This feels kind of weird since it's not very explicit in showing that it
is only meant for toggling the `is_paid` field. It feels more suited for
GET stuff as a filter or something similar.

This patch makes it so you have to explicitly include `?is_paid=1` or
`?is_paid=0`. This will also make it easier in the future to implement
editing other fields of the debt by accessing the endpoint without any
query parameters.