Avoid fetching from database when switching users
It seems like constantly fetching from the database whenever a user
looks through the list of users is kind of excessive. So, we just never
get new data unless the user decides to refresh the page.

This patch also attempts to improve responsiveness by not `await`ing any
`fetch()` calls when we don't need the result.

Error handling is also slightly improved with new `catch` statements.