Let users access their group by its `short_name`
A group in the `groups` collection in the database can now (optionally)
contain the `short_name` field.

This can be used as a shortcut URL for people who don't want to deal
with the super-long group IDs.

For example, let's say we have the following group:

    {
      _id: ObjectId("1234abcd"),
      name: "Foo",
      short_name: "foo",
      user_list: [
        // ...
      ]
    }

We were able to access this group at `/1234abcd`, but now we can also
access it at `/foo`.